requirements.txt
absl-py==0.9.0
alabaster==0.7.12
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
astroid==2.3.3
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
black==19.10b0
bleach==3.1.0
c3d==0.3.0
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.3
django-crispy-forms==1.7.2
docutils==0.15.2
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.8
image==1.5.27
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
json5==0.9.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.2.6
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lesscpy==0.13.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
modern-robotics==1.1.0
more-itertools==7.2.0
mpmath==1.1.0
mypy==0.740
mypy-extensions==0.4.3
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.6.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.4
pyrsistent==0.15.4
pytest==5.3.0
python-dateutil==2.8.0
pytz==2018.7
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.11
sympy==1.4
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
